+++ src/component/Breakfast/BreakFastShrimp.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { Container, Card, CardText, CardTitle, Button } from "reactstrap";
 import bf_shrimp from "../../imgbreakfast/กุ้งชุบแป้งทอด.png";
+import axios from "axios";
 export default class BreakFastShrimp extends Component {
   state = {
     menu_name: "กุ้งชุบแป้งทอด",
@@ -8,23 +9,42 @@ export default class BreakFastShrimp extends Component {
     menu_price: 40,
   };
 
-  sentOrder = (e) => {
+  sentOrder = async (e) => {
     e.preventDefault();
+    let token = localStorage.getItem("token");
+    if (token) {
+      if (token !== null) {
+        this.setState({ check: "login" });
+        var oldItems = JSON.parse(localStorage.getItem("order")) || [];
 
-    var oldItems = JSON.parse(localStorage.getItem("order")) || [];
+        console.log("sent !");
 
-    console.log("sent !");
+        const newData = {
+          menu_name: this.state.menu_name,
+          menu_value: 1,
+          menu_price: this.state.menu_price,
+        };
 
-    const newData = {
-      menu_name: this.state.menu_name,
-      menu_value: 1,
-      menu_price: this.state.menu_price,
-    };
-    oldItems.push(newData);
-    localStorage.setItem("order", JSON.stringify(oldItems));
-    console.log(" data : ", newData);
-    this.setState({ menu_value: 0 });
-    this.setState({ remark: "" });
+        const isHaveMenuAlready = oldItems.find(
+          (menu) => menu.menu_name === this.state.menu_name
+        );
+
+        if (!isHaveMenuAlready) {
+          oldItems.push(newData);
+          localStorage.setItem("order", JSON.stringify(oldItems));
+          console.log(" data : ", newData);
+          this.setState({ menu_value: 0 });
+          this.setState({ remark: "" });
+          window.location.reload();
+        }
+      }
+      let res = await axios.post(`http://localhost:3001/api/users/me`, {
+        token,
+      });
+      this.setState({ user: res.data });
+    } else {
+      window.location.href = "/login";
+    }
   };
 
   handleInputChange = (e) => {
@@ -38,10 +58,10 @@ export default class BreakFastShrimp extends Component {
       <div>
         <Container>
           <div className="form-menu">
-            <Card body>
+            <Card body className="border border-info">
               <div className="pic_buk row justify-content-md-center">
                 <img
-                  className="img_bfshrimp img-fluid"
+                  className="img_bfshrimp img-fluid "
                   src={bf_shrimp}
                   alt=""
                 />
